Remove unnecessary fromIntegral calls

+++ libraries/base/GHC/Conc/IO.hs
@@ -107,7 +107,7 @@ threadDelay time
   | threaded  = Event.threadDelay time
 #endif
   | otherwise = IO $ \s ->
-        case fromIntegral time of { I# time# ->
+        case time of { I# time# ->
         case delay# time# s of { s' -> (# s', () #)
         }}
